+++ tools/integration_tests/explicit_dir/explicit_dir_test.go
@@ -16,13 +16,20 @@
 package explicit_dir_test
 
 import (
+	"os"
 	"testing"
 
+	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/setup"
 	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/setup/implicit_and_explicit_dir_setup"
 )
 
 func TestMain(m *testing.M) {
+	setup.ParseSetUpFlags()
+
 	flags := [][]string{{"--implicit-dirs=false"}}
 
-	implicit_and_explicit_dir_setup.RunTestsForImplicitDirAndExplicitDir(flags, m)
+	successCode := implicit_and_explicit_dir_setup.RunTestsForImplicitDirAndExplicitDir(flags, m)
+
+	setup.RemoveBinFileCopiedForTesting()
+	os.Exit(successCode)
 }
