+++ tools/integration_tests/run_tests_mounted_directory.sh
