+++ tools/integration_tests/util/mounting/only_dir_mounting/only_dir_mounting.go
@@ -77,7 +77,8 @@ func executeTestsForOnlyDirMounting(flags [][]string, m *testing.M) (successCode
 	successCode = mountGcsFuseForFlagsAndExecuteTests(flags, DirectoryInTestBucket, m)
 
 	// Clean the bucket after testing.
-	//setup.RunScriptForTestData("../util/mounting/only_dir_mounting/testdata/delete_objects.sh", setup.TestBucket())
+	setup.RunScriptForTestData("../util/mounting/only_dir_mounting/testdata/delete_objects.sh", setup.TestBucket())
+
 	// Reset onlyDirMounted value to empty string after only dir mount tests are done.
 	setup.SetOnlyDirMounted("")
 	return
