+++ tools/integration_tests/util/setup/implicit_and_explicit_dir_setup/implicit_and_explicit_dir_setup.go
@@ -41,9 +41,7 @@ const SecondFileInExplicitDirectory = "fileInExplicitDir2"
 const FileInImplicitDirectory = "fileInImplicitDir1"
 const FileInImplicitSubDirectory = "fileInImplicitDir2"
 
-func RunTestsForImplicitDirAndExplicitDir(flags [][]string, m *testing.M) {
-	setup.ParseSetUpFlags()
-
+func RunTestsForImplicitDirAndExplicitDir(flags [][]string, m *testing.M) int {
 	setup.ExitWithFailureIfBothTestBucketAndMountedDirectoryFlagsAreNotSet()
 
 	if setup.TestBucket() == "" && setup.MountedDirectory() != "" {
@@ -62,10 +60,7 @@ func RunTestsForImplicitDirAndExplicitDir(flags [][]string, m *testing.M) {
 	if successCode == 0 {
 		successCode = persistent_mounting.RunTests(flags, m)
 	}
-
-	setup.RemoveBinFileCopiedForTesting()
-
-	os.Exit(successCode)
+	return successCode
 }
 
 func RemoveAndCheckIfDirIsDeleted(dirPath string, dirName string, t *testing.T) {
